rename resource because of conflict in unittest lib added unittest settings for VSCode
