+++ karabo/karabo_resource.py
